Create Release_Prison
Imagine you are in a prison with many prison cells. The cells can be locked or unlocked. We represent the cells as an array of e's and 1's, where o represents an unlocked cell and 1 represents a locked cell. Each cell has exactly 1 prisoner. You are the prisoner in the leftmost cell. So, if the configuration is 10, it means there are 3 cells: you are in an unlocked cell, the next cell is locked, and the cell after that is also unlocked.

If your cell is unlocked, you can walk down the cells in order and release different prisoners. The cell must be unlocked (e) for you to be able to release the prisoner. Once you unlock a cell, all cells will flip from locked to unlocked, and vice-versa. So, if the state is a 10, it will become 1 0 1, for example.

Once you release a prisoner, you cannot move back to the previous cell. You can only move forward. Your task is to return the number of prisoners that you can release, num_release.